# performance improvment via .bin store cost next to state_index