--- a/.gitignore
+++ b/.gitignore
@@ -104,3 +104,5 @@ ida_search.dSYM
 ida_search_via_graph
 lookup-tables
 my-pt-states.txt
+gmon.out
+prof_output