--- a/.gitignore
+++ b/.gitignore
@@ -106,3 +106,4 @@ lookup-tables
 my-pt-states.txt
 gmon.out
 prof_output
+NOTES.txt