--- a/rubikscubennnsolver/LookupTableIDAViaGraph.py
+++ b/rubikscubennnsolver/LookupTableIDAViaGraph.py
@@ -71,6 +71,7 @@ class LookupTableIDAViaGraph(LookupTable):
         pt1_state_max: int = None,
         pt2_state_max: int = None,
         centers_only: bool = False,
+        use_uthash: bool = False,
     ):
         LookupTable.__init__(self, parent, filename, state_target, linecount, max_depth, filesize)
         self.recolor_positions = []
@@ -84,6 +85,7 @@ class LookupTableIDAViaGraph(LookupTable):
         self.main_table_max_depth = main_table_max_depth
         self.main_table_prune_tables = main_table_prune_tables
         self.centers_only = centers_only
+        self.use_uthash = use_uthash
 
         if perfect_hash01_filename:
             self.perfect_hash01_filename = "lookup-tables/" + perfect_hash01_filename
@@ -302,6 +304,9 @@ class LookupTableIDAViaGraph(LookupTable):
 
         if self.centers_only:
             cmd.append("--centers-only")
+
+        if self.use_uthash:
+            cmd.append("--uthash")
 
         cmd.append("--legal-moves")
         cmd.append(",".join(self.all_moves))