… NOTES.txt
… rubikscubennnsolver/RubiksCube222.py
… rubikscubennnsolver/RubiksCube333.py
… rubikscubennnsolver/RubiksCube444.py
… rubikscubennnsolver/RubiksCube444Misc.py
… rubikscubennnsolver/RubiksCube555.py
… rubikscubennnsolver/RubiksCube666.py
… rubikscubennnsolver/RubiksCube777.py
… rubikscubennnsolver/RubiksCubeNNNEven.py
… rubikscubennnsolver/RubiksCubeNNNEvenEdges.py
… rubikscubennnsolver/RubiksCubeNNNOdd.py
… rubikscubennnsolver/RubiksCubeNNNOddEdges.py
… rubikscubennnsolver/__init__.py
… rubikscubennnsolver/ida_search.c
… rubikscubennnsolver/ida_search_444.c
… rubikscubennnsolver/ida_search_444.h
… rubikscubennnsolver/ida_search_666.c
… rubikscubennnsolver/ida_search_666.h
… rubikscubennnsolver/ida_search_777.c
… rubikscubennnsolver/ida_search_777.h
… rubikscubennnsolver/ida_search_core.c
… rubikscubennnsolver/ida_search_core.h
… rubikscubennnsolver/ida_search_via_graph.c
… rubikscubennnsolver/misc.py
… utils/build-444-phase3-stats.py
… utils/build-555-LR-centers-stage-stats.py
… utils/build-777-LR-oblique-pairing-stats.py
… utils/build-777-UD-oblique-pairing-stats.py
… utils/build-neural-network-stats.py
… utils/build-test-cases.py