A huge performance gain was made by rebuilding the .bin files so that for a given step we not only stored the state_index
of the next node but the cost of the next node. This allows the state_index and the cost of the next node to both be in
the CPU cache which drastically reduces the amount of time to lookup the cost.

before
```
./ida_search_via_graph --prune-table-0-filename lookup-tables/lookup-table-5x5x5-step11-LR-centers-stage-t-center-only.bin --prune-table-0-state 218536 --prune-table-1-filename lookup-tables/lookup-table-5x5x5-step12-LR-centers-stage-x-center-only.bin --prune-table-1-state 86325 --centers-only --solution-count 10 --legal-moves "U,U',U2,Uw,Uw',Uw2,L,L',L2,Lw,Lw',Lw2,F,F',F2,Fw,Fw',Fw2,R,R',R2,Rw,Rw',Rw2,B,B',B2,Bw,Bw',Bw2,D,D',D2,Dw,Dw',Dw2"
[2021-09-03T16:54:46.253] IDA threshold 7, explored 717 nodes, took 0.000s, 0 nodes-per-sec
[2021-09-03T16:54:46.255] IDA threshold 8, explored 22,766 nodes, took 0.002s, 11,383,000 nodes-per-sec
[2021-09-03T16:54:46.309] IDA threshold 9, explored 525,053 nodes, took 0.053s, 9,906,660 nodes-per-sec
[2021-09-03T16:54:47.605] IDA threshold 10, explored 12,616,044 nodes, took 1.295s, 9,742,119 nodes-per-sec
[2021-09-03T16:54:47.956] IDA count 3,419,526, f_cost 11 vs threshold 11 (cost_to_here 11, cost_to_goal 0)
SOLUTION (11 steps): D2 Rw' D2 Fw Rw' B Uw' Lw' D Dw Bw' 
SOLUTION (11 steps): D2 Rw' D2 Fw Rw' B Uw' Lw' D Dw Bw 
SOLUTION (11 steps): D2 Fw R2 Dw2 Bw' Uw' Bw Lw2 U Dw2 Bw' 
SOLUTION (11 steps): D2 Fw R2 Dw2 Bw' Uw' Bw Lw2 U Dw2 Bw 
SOLUTION (11 steps): D2 Fw R Dw2 Bw' Uw' Bw Uw2 Fw2 D Fw' 
SOLUTION (11 steps): D2 Fw R Dw2 Bw' Uw' Bw Uw2 Fw2 D Fw 
SOLUTION (11 steps): D2 Lw' R' Dw Lw' Bw U Fw2 Lw' Fw Dw' 
SOLUTION (11 steps): D2 Lw' R' Dw Lw' Bw U Fw2 Lw' Fw Dw 
SOLUTION (11 steps): D' Bw Rw2 U2 D' Fw B2 Uw' Bw2 Dw2 Bw' 
SOLUTION (11 steps): D' Bw Rw2 U2 D' Fw B2 Uw' Bw2 Dw2 Bw 
[2021-09-03T16:54:48.591] IDA threshold 11, explored 8,988,659 nodes, took 0.987s, 9,107,051 nodes-per-sec
[2021-09-03T16:54:48.591] IDA found solution, explored 22,153,239 total nodes, took 2.338s, 9,475,295 nodes-per-sec
```

after
```
./ida_search_via_graph --prune-table-0-filename lookup-tables/lookup-table-5x5x5-step11-LR-centers-stage-t-center-only.bin --prune-table-0-state 218536 --prune-table-1-filename lookup-tables/lookup-table-5x5x5-step12-LR-centers-stage-x-center-only.bin --prune-table-1-state 86325 --centers-only --solution-count 10 --legal-moves "U,U',U2,Uw,Uw',Uw2,L,L',L2,Lw,Lw',Lw2,F,F',F2,Fw,Fw',Fw2,R,R',R2,Rw,Rw',Rw2,B,B',B2,Bw,Bw',Bw2,D,D',D2,Dw,Dw',Dw2"
[2021-09-09T09:50:54.858] pt0_state 218536, pt1_state 86325, pt2_state 0, pt3_state 0, pt4_state 0
[2021-09-09T09:50:54.858] IDA threshold 7, explored 720 nodes, took 0.000s, 0 nodes-per-sec
[2021-09-09T09:50:54.859] IDA threshold 8, explored 23,067 nodes, took 0.000s, 0 nodes-per-sec
[2021-09-09T09:50:54.866] IDA threshold 9, explored 547,275 nodes, took 0.007s, 78,182,144 nodes-per-sec
[2021-09-09T09:50:55.064] IDA threshold 10, explored 13,310,568 nodes, took 0.198s, 67,225,096 nodes-per-sec
[2021-09-09T09:50:55.301] IDA count 16,672,377, f_cost 11 vs threshold 11 (cost_to_here 11, cost_to_goal 0)
SOLUTION (11 steps): D2 Rw' D2 Fw Rw' B Uw' Lw' D Dw Bw' 
SOLUTION (11 steps): D2 Rw' D2 Fw Rw' B Uw' Lw' D Dw Bw 
SOLUTION (11 steps): D2 Fw R2 Dw2 Bw' Uw' Bw Lw2 U Dw2 Bw' 
SOLUTION (11 steps): D2 Fw R2 Dw2 Bw' Uw' Bw Lw2 U Dw2 Bw 
SOLUTION (11 steps): D2 Fw R Dw2 Bw' Uw' Bw Uw2 Fw2 D Fw' 
SOLUTION (11 steps): D2 Fw R Dw2 Bw' Uw' Bw Uw2 Fw2 D Fw 
SOLUTION (11 steps): D2 Lw' R' Dw Lw' Bw U Fw2 Lw' Fw Dw' 
SOLUTION (11 steps): D2 Lw' R' Dw Lw' Bw U Fw2 Lw' Fw Dw 
SOLUTION (11 steps): D' Bw Rw2 U2 D' Fw B2 Uw' Bw2 Dw2 Bw' 
SOLUTION (11 steps): D' Bw Rw2 U2 D' Fw B2 Uw' Bw2 Dw2 Bw 
[2021-09-09T09:50:55.413] IDA threshold 11, explored 23,446,560 nodes, took 0.348s, 67,375,168 nodes-per-sec
[2021-09-09T09:50:55.413] IDA found solution, explored 37,328,190 total nodes, took 0.555s, 67,258,000 nodes-per-sec
```

5x5x5 solutions were improved a little via:
- try x and y rotations when staging centers
- attempting to pair many more wing_str_combos for the first four edges

The 5x5x5 improvements in turn reduced the move count for 6x6x6 and larger cubes.